[MINOR] fix a few remaining printf-like formats on 64-bit platforms
Mainly two sizeof() returning size_t which is not the size of an int
on 64-bit platforms.
(cherry picked from commit 5e4a6f13f4b3c3de0e533284d5fabd9b4bbe00cc)

+++ src/standard.c
@@ -92,7 +92,7 @@ struct sockaddr_un *str2sun(const char *str)
 	strsz = strlen(str) + 1;
 	if (strsz > sizeof(su.sun_path)) {
 		Alert("Socket path '%s' too long (max %d)\n",
-			str, sizeof(su.sun_path) - 1);
+		      str, (int)sizeof(su.sun_path) - 1);
 	} else {
 		su.sun_family = AF_UNIX;
 		memcpy(su.sun_path, str, strsz);
